ab strassen - still needs work

+++ step6/dgemm/my_dgemm_strassen_abc.c
@@ -54,7 +54,7 @@ inline void packA_add_mcxkc_d_str_abc(
         int    k,
         double *XAA,
         double *XAB,
-        int    gamma,
+        double gamma,
         int    ldXA,
         int    offseta,
         double *packA
@@ -127,7 +127,7 @@ inline void packB_add_kcxnc_d_str_abc(
         int    k,
         double *XBA,
         double *XBB,
-        int    gamma,
+        double gamma,
         int    ldXB, // ldXB is the original k
         int    offsetb,
         double *packB
@@ -198,8 +198,8 @@ void bl_macro_kernel_str_abc(
         double *CA,
         double *CB,
         int    ldc,
-        int    gammaCA,
-        int    gammaCB
+        double gammaCA,
+        double gammaCB
         )
 {
     int bl_ic_nt;
@@ -271,17 +271,17 @@ void bl_dgemm_str_abc(
         double *XAB,
         double *packA,
         int    lda,
-        int    gammaA,
+        double gammaA,
         double *XBA,
         double *XBB,
         double *packB,
         int    ldb,
-        int    gammaB,
+        double gammaB,
         double *CA,        // must be aligned
         double *CB,
         int    ldc,        // ldc must also be aligned
-        int    gammaCA,
-        int    gammaCB,
+        double gammaCA,
+        double gammaCB,
         int    bl_ic_nt
         )
 {
